Clean up parquet file versioning logic.
